Transform data according to the requirements
