--- services/market/shared/market/sources/constants/exchangeratesapi.js
@@ -1,69 +1,0 @@
-/*
- * LiskHQ/lisk-service
- * Copyright © 2021 Lisk Foundation
- *
- * See the LICENSE file at the top-level directory of this distribution
- * for licensing information.
- *
- * Unless otherwise agreed in a custom licensing agreement with the Lisk Foundation,
- * no part of this software, including this file, may be copied, modified,
- * propagated, or distributed except according to the terms contained in the
- * LICENSE file.
- *
- * Removal or modification of this copyright notice is prohibited.
- *
- */
-const symbolMap = {
-    EUR_USD: 'EURUSD',
-    EUR_GBP: 'EURGBP',
-    EUR_CHF: 'EURCHF',
-    EUR_BTC: 'EURBTC',
-    EUR_RUB: 'EURRUB',
-    EUR_JPY: 'EURJPY',
-
-    USD_EUR: 'USDEUR',
-    USD_GBP: 'USDGBP',
-    USD_CHF: 'USDCHF',
-    USD_BTC: 'USDBTC',
-    USD_RUB: 'USDRUB',
-    USD_JPY: 'USDJPY',
-
-    GBP_USD: 'GBPUSD',
-    GBP_EUR: 'GBPEUR',
-    GBP_CHF: 'GBPCHF',
-    GBP_BTC: 'GBPBTC',
-    GBP_RUB: 'GBPRUB',
-    GBP_JPY: 'GBPJPY',
-
-    CHF_USD: 'CHFUSD',
-    CHF_GBP: 'CHFGBP',
-    CHF_EUR: 'CHFEUR',
-    CHF_BTC: 'CHFBTC',
-    CHF_RUB: 'CHFRUB',
-    CHF_JPY: 'CHFJPY',
-
-    BTC_USD: 'BTCUSD',
-    BTC_GBP: 'BTCGBP',
-    BTC_CHF: 'BTCCHF',
-    BTC_EUR: 'BTCEUR',
-    BTC_RUB: 'BTCRUB',
-    BTC_JPY: 'BTCJPY',
-
-    RUB_USD: 'RUBUSD',
-    RUB_GBP: 'RUBGBP',
-    RUB_CHF: 'RUBCHF',
-    RUB_BTC: 'RUBBTC',
-    RUB_EUR: 'RUBEUR',
-    RUB_JPY: 'RUBJPY',
-
-    JPY_USD: 'JPYUSD',
-    JPY_GBP: 'JPYGBP',
-    JPY_CHF: 'JPYCHF',
-    JPY_BTC: 'JPYBTC',
-    JPY_RUB: 'JPYRUB',
-    JPY_EUR: 'JPYEUR',
-};
-
-module.exports = {
-    symbolMap,
-};
